Analytical results for variance estimation of linear-linear specifications
The goal here is to use the results we get out of models of the form Y = a'x + b'z + e to check the output of vcov and similar when first fitting Y = a'x + e_1 then (Y - \hat a'x) = b'z + e_2. The current results use the classical approach of estimating variance of estimators as \hat sigma^2 [(X Z)'(X Z)]^{-1} and express these quantities in terms of X, Z, and Y.

What is currently missing is the extension to the estimating equations framework. There are some notes in the vignette about where to go to get some of these materials.