These were some fixes I tried that I don't think I need, but want to save
Specifically, this was from an approach where instead of trying to address scoping head-on, I instead was making helper functions more robust to scoping issues. This 'worked', but I worry that it will fail in unexpected ways in more scenarios. I generally have better solutions, but saving this code here in case I need to return to it.