Replace scrambled version with clean version from branch
The versions of estfun_DA.tex on main and on the absorb
branch had some differences in how they represented weighting.
Unfortunately the version on main became scrambled in the midst
of an unrelated merge. Lacking the time to figure out how to revert
it to its prior state, I'll overwrite it with the absorb version,
bringing the 2 up to date with one another.