Numeric-stability rescaling substantive update proposal
The proposal, to fold into DA estfuns a combination of
\tilde{A}_{22} (n^{-1} scaling) and A_{21} (n_{Q}^{-1}
scaling; no tilde), is reflected only in the Numerical
Stability segment of sec:conc-vs-pract.  If we go with
this, other sections of the spec remain to be updated.
(Various tokens of (n^{1/2}/n_{C}^{1/2}) would be replaced
by n_Q/(n_C^{1/2}n^{1/2}).)