BUG: in sec:realization-goal ff, estfun.DA's incorrectly take sqrt of scaling factor
This commit corrects the exponent in eq:15, which incorrectly had a square root on
the scaling factor.  It also edits surrounding displays and language, for clarity, and
propagates the exponent fix into displays eq:10, eq:24 and eq:12.  In the case of
eq:12, there may have been additional issues with scaling factors having to do with
scale of M_21.

Below sec:realization-goal, several places the commit at the same time
corrects the exponent error and adjust expressions for the `estfun.DA()`
return value to assume it will ultimately be pre-multiplied by \hat{\tilde{A}}_{22},
not \hat{A}_{22}. First (n/n_C)^{1/2} becomes (n/n_C); then it becomes (n_Q/n_C).
(This describes changes in eq:26, eq:25, and  eq:19.)