Adjust Rd for `[`-methods to try to dodge warning
On my local machine this circumvents a warning that's causing
the CI to error out, so hopefully it will fix that problem. If so
then it's a bug on the checker's part, AFAICS based on Writing R
Extensions Sec 2.1.3.  If/when that issue is fixed, this commit
can be reverted.